Integrate React Native Paper UI library
Based on Google's Material UI
