+++ package.json
@@ -20,6 +20,8 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.72.4",
+    "react-native-paper": "^5.10.5",
+    "react-native-safe-area-context": "^4.7.2",
     "react-native-web": "~0.19.6"
   },
   "devDependencies": {
